server/python/requirements.txt
Flask
PyPDF2
python-dotenv
google-generativeai
langchain-core
